feat(AddFriends): splits out addfriends search & button to new component, hooks up feature to create event page
